Move grpc_query to module root
